feat(data-model): 将集群ID参数类型统一更改为Long，提升数据一致性
- 更新多个类中的集群ID字段类型，从Integer更改为Long，以确保数据模型的一致性与兼容性。
- 涉及的类包括CacheUtils、BaseCommand、ClusterCommand、DispatcherHostAgentCommand等，确保所有相关逻辑均能正确处理集群ID。

这些改动显著增强了数据模型的健壮性与一致性，确保了整体设计的一致性。

+++ datasophon-common/src/main/java/com/datasophon/common/cache/CacheUtils.java
@@ -180,7 +180,7 @@ public class CacheUtils {
      * @param ip        主机IP
      * @param hostInfo  主机信息对象
      */
-    public static void putHostInfo(Integer clusterId, String ip, HostInfo hostInfo) {
+    public static void putHostInfo(Long clusterId, String ip, HostInfo hostInfo) {
         if (clusterId == null || StringUtils.isBlank(ip) || hostInfo == null) {
             logger.warn("更新主机缓存参数无效: clusterId={}, ip={}", clusterId, ip);
             return;
