+++ datasophon-common/src/main/java/com/datasophon/common/command/GenerateRackPropCommand.java
@@ -23,6 +23,6 @@ import lombok.Data;
 @Data
 public class GenerateRackPropCommand implements Serializable {
 
-    private Integer clusterId;
+    private Long clusterId;
 
 }
